+++ source/p2p_app/chat/mod.rs
@@ -2,3 +2,4 @@
 
 pub mod cli;
 pub mod gui;
+pub mod login_common;